Move Instruction into its own module
