+++ programs/budget_api/src/budget_instruction.rs
@@ -2,8 +2,8 @@ use crate::budget_expr::BudgetExpr;
 use crate::id;
 use chrono::prelude::{DateTime, Utc};
 use serde_derive::{Deserialize, Serialize};
+use solana_sdk::instruction::{AccountMeta, Instruction};
 use solana_sdk::pubkey::Pubkey;
-use solana_sdk::transaction::{AccountMeta, Instruction};
 
 /// A smart contract.
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
