+++ programs/config_api/src/config_instruction.rs
@@ -1,8 +1,8 @@
 use crate::id;
 use crate::ConfigState;
+use solana_sdk::instruction::{AccountMeta, Instruction};
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::system_instruction::SystemInstruction;
-use solana_sdk::transaction::{AccountMeta, Instruction};
 
 pub struct ConfigInstruction {}
