+++ programs/config_api/src/config_processor.rs
@@ -2,8 +2,8 @@
 
 use log::*;
 use solana_sdk::account::KeyedAccount;
+use solana_sdk::instruction::InstructionError;
 use solana_sdk::pubkey::Pubkey;
-use solana_sdk::transaction::InstructionError;
 
 pub fn process_instruction(
     _program_id: &Pubkey,
