+++ programs/noop_program/src/lib.rs
@@ -1,8 +1,8 @@
 use log::*;
 use solana_sdk::account::KeyedAccount;
+use solana_sdk::instruction::InstructionError;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::solana_entrypoint;
-use solana_sdk::transaction::InstructionError;
 
 solana_entrypoint!(entrypoint);
 fn entrypoint(
