+++ programs/token_api/src/token_processor.rs
@@ -2,8 +2,8 @@ use crate::token_state::TokenState;
 use bincode::serialize;
 use log::*;
 use solana_sdk::account::KeyedAccount;
+use solana_sdk::instruction::InstructionError;
 use solana_sdk::pubkey::Pubkey;
-use solana_sdk::transaction::InstructionError;
 
 pub fn process_instruction(
     program_id: &Pubkey,
