+++ runtime/src/native_loader.rs
@@ -6,10 +6,10 @@ use libloading::os::unix::*;
 use libloading::os::windows::*;
 use log::*;
 use solana_sdk::account::KeyedAccount;
+use solana_sdk::instruction::InstructionError;
 use solana_sdk::loader_instruction::LoaderInstruction;
 use solana_sdk::native_program;
 use solana_sdk::pubkey::Pubkey;
-use solana_sdk::transaction::InstructionError;
 use std::env;
 use std::path::PathBuf;
 use std::str;
